backend/pyproject.toml
[project]
name = "agent-topic-lab-backend"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "claude-agent-sdk>=0.1.29",
    "anthropic>=0.39.0",
    "openai>=1.0.0",
]

[tool.setuptools]
packages = ["app"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.26.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
testpaths = ["tests"]
